the result of the 14 sprint

+++ controllers/cards.js
@@ -1,16 +1,20 @@
-const validationError = require('mongoose').Error.ValidationError;
-const castError = require('mongoose').Error.CastError;
 const Card = require('../models/card');
+const {
+  NOT_FOUND,
+  UNAUTHORIZED,
+} = require('../utils/constants');
 
-const INTERNAL_SERVER_ERROR = 500;
-const BAD_REQUEST = 400;
-const NOT_FOUND = 404;
+const { validateCard, validateCardId } = require('../middlewares/validation');
 
-module.exports.getCards = (req, res) => Card.find({})
-  .then((cards) => res.status(200).send({ data: cards }))
-  .catch(() => res.status(INTERNAL_SERVER_ERROR).send({ message: 'Внутренняя ошибка сервера' }));
+module.exports.getCards = (req, res, next) => {
+  Card.find({})
+    .then((cards) => res.status(200).send({ data: cards }))
+    .catch((err) => {
+      next(err);
+    });
+};
 
-module.exports.postCard = (req, res) => {
+module.exports.postCard = (req, res, next) => {
   const { name, link } = req.body;
 
   const cardData = {
@@ -24,30 +28,38 @@ module.exports.postCard = (req, res) => {
       res.status(201).send({ data: card });
     })
     .catch((err) => {
-      if (err instanceof validationError) {
-        res.status(BAD_REQUEST).send({ message: 'Ошибка при валидации' });
-      } else {
-        res.status(INTERNAL_SERVER_ERROR).send({ message: 'Внутренняя ошибка сервера' });
-      }
+      next(err);
     });
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const { cardId } = req.params;
-  Card.findByIdAndRemove(cardId)
+  const userId = req.user._id;
+
+  Card.findById(cardId)
     .then((card) => {
       if (!card) {
-        res.status(NOT_FOUND).send({ message: 'Карточка с таким id не найдена' });
-      } else {
-        res.status(200).send(card);
+        return res.status(NOT_FOUND).send({ message: 'Карточка с таким id не найдена' });
       }
+
+      if (card.owner.toString() !== userId) {
+        return res.status(UNAUTHORIZED).send({ message: 'Нельзя удалить чужую карточку' });
+      }
+
+      return Card.findByIdAndRemove(cardId)
+        .then((removedCard) => {
+          if (!removedCard) {
+            res.status(NOT_FOUND).send({ message: 'Карточка с таким id не найдена' });
+          } else {
+            res.status(200).send(removedCard);
+          }
+        })
+        .catch((err) => {
+          next(err);
+        });
     })
     .catch((err) => {
-      if (err instanceof castError) {
-        res.status(BAD_REQUEST).send({ message: 'Передан некорректный id карточки' });
-      } else {
-        res.status(INTERNAL_SERVER_ERROR).send({ message: 'Внутренняя ошибка сервера' });
-      }
+      next(err);
     });
 };
 
@@ -62,11 +74,8 @@ const handleCardOperation = (req, res, operation) => {
       }
     })
     .catch((err) => {
-      if (err instanceof castError) {
-        res.status(BAD_REQUEST).send({ message: 'Передан некорректный id карточки' });
-      } else {
-        res.status(INTERNAL_SERVER_ERROR).send({ message: 'Внутренняя ошибка сервера' });
-      }
+      // eslint-disable-next-line no-undef
+      next(err);
     });
 };
 
@@ -86,4 +95,9 @@ module.exports.dislikeCard = (req, res) => {
     { new: true },
   ).populate('owner');
   handleCardOperation(req, res, operation);
+};
+
+module.exports = {
+  validateCard,
+  validateCardId,
 };
